alterations done.

+++ lib/features/device_details/view/device_detail_screens/model/pf_data_model.dart
@@ -1,12 +1,12 @@
 class Pf {
   String? unit;
-  int? value;
+  double? value;
 
   Pf({this.unit, this.value});
 
   Pf.fromJson(Map<String, dynamic> json) {
     unit = json['unit'];
-    value = (json['value'] is double) ? json['value'].toInt() : json['value'];
+    value = (json['value'] is double) ? json['value'].toDouble() : json['value'];
   }
 
   Map<String, dynamic> toJson() {
